fix require error, signal_level regexp error & dir restructuring
