+++ lib/iwlist.js
@@ -16,7 +16,7 @@ function parseIwlist(str) {
         // 'encryption_key' : /Encryption key:(.*)/,
         // 'bitrates' : /Bit Rates:(.*)/,
         // 'quality' : /Quality(?:=|\:)([^\s]+)/,
-        'signal_level' : /Signal level(?:=|\:)([\w]+)/
+        'signal_level' : /Signal level(?:=|\:)([-\w]+)/
     };
 
     for (var i=0,l=out.length; i<l; i++) {
