+++ lib/wifiscanner.js
@@ -1,6 +1,6 @@
 var fs = require('fs');
-var airport = require('lib/airport');
-var iwlist = require('lib/iwlist');
+var airport = require('./airport');
+var iwlist = require('./iwlist');
 
 function scan(callback) {
     fs.exists(airport.utility, function (exists) {
